Generate a baseline compilation report
This commit introduces a baseline compilation report. The report
contains the events that can occur when modifying a mainline catalog.

* Resource addition
* Parameter addition
* Parameter overwrites

The report is saved using an indirector.  Currently, there is only a
JSON indirector that saves the file out to a json file located at
$client_datadir/baselinereport/$hostname_baselinereport.json

In the future, more indirectors can be added to extend where the
baseline report can be stored.